explicitly skip tests for platforms that not support unix sockets
